En esta version, se hace el promedio per capita... o sea que el puntaje total obtenido por una estrategia se divide por la cantidad de individuos de dicha estrategia.

+++ src/main/java/ar/fi/uba/graficos/Grafico.java
@@ -33,7 +33,7 @@ public class Grafico {
         seriesSegunEstrategiaCantidad.get(estrategia).add(generacion, cantIndividuos);
     }
     
-    public void agregarDatosPuntaje(String estrategia, Integer generacion, Long puntaje) {
+    public void agregarDatosPuntaje(String estrategia, Integer generacion, Double puntaje) {
         seriesSegunEstrategiaPuntaje.get(estrategia).add(generacion, puntaje);
     }
